Rename: loadsuppliers -> companies
