+++ jarbas/core/tests/test_companies_command.py
@@ -4,7 +4,7 @@ from unittest.mock import patch
 
 from django.test import TestCase
 
-from jarbas.core.management.commands.loadsuppliers import Command
+from jarbas.core.management.commands.companies import Command
 from jarbas.core.models import Activity, Supplier
 from jarbas.core.tests import sample_supplier_data
 
@@ -61,11 +61,11 @@ class TestCreate(TestCommand):
         self.assertEqual(1, len(main))
         self.assertEqual(99, len(secondaries))
 
-    @patch('jarbas.core.management.commands.loadsuppliers.lzma')
-    @patch('jarbas.core.management.commands.loadsuppliers.csv.DictReader')
-    @patch('jarbas.core.management.commands.loadsuppliers.Command.save_activities')
-    @patch('jarbas.core.management.commands.loadsuppliers.Command.serialize')
-    @patch('jarbas.core.management.commands.loadsuppliers.Command.print_count')
+    @patch('jarbas.core.management.commands.companies.lzma')
+    @patch('jarbas.core.management.commands.companies.csv.DictReader')
+    @patch('jarbas.core.management.commands.companies.Command.save_activities')
+    @patch('jarbas.core.management.commands.companies.Command.serialize')
+    @patch('jarbas.core.management.commands.companies.Command.print_count')
     @patch.object(Supplier.objects, 'create')
     def test_save_suppliers(self, create, print_count, serialize, save_activities, rows, lzma):
         self.command.count = 0
@@ -82,10 +82,10 @@ class TestCreate(TestCommand):
 
 class TestConventionMethods(TestCommand):
 
-    @patch('jarbas.core.management.commands.loadsuppliers.print')
-    @patch('jarbas.core.management.commands.loadsuppliers.LoadCommand.drop_all')
-    @patch('jarbas.core.management.commands.loadsuppliers.Command.save_suppliers')
-    @patch('jarbas.core.management.commands.loadsuppliers.Command.print_count')
+    @patch('jarbas.core.management.commands.companies.print')
+    @patch('jarbas.core.management.commands.companies.LoadCommand.drop_all')
+    @patch('jarbas.core.management.commands.companies.Command.save_suppliers')
+    @patch('jarbas.core.management.commands.companies.Command.print_count')
     def test_handler_without_options(self, print_count, save_suppliers, drop_all, print_):
         print_count.return_value = 0
         self.command.handle(dataset='companies.xz')
@@ -95,10 +95,10 @@ class TestConventionMethods(TestCommand):
         self.assertEqual('companies.xz', self.command.path)
         drop_all.assert_not_called()
 
-    @patch('jarbas.core.management.commands.loadsuppliers.print')
-    @patch('jarbas.core.management.commands.loadsuppliers.Command.drop_all')
-    @patch('jarbas.core.management.commands.loadsuppliers.Command.save_suppliers')
-    @patch('jarbas.core.management.commands.loadsuppliers.Command.print_count')
+    @patch('jarbas.core.management.commands.companies.print')
+    @patch('jarbas.core.management.commands.companies.Command.drop_all')
+    @patch('jarbas.core.management.commands.companies.Command.save_suppliers')
+    @patch('jarbas.core.management.commands.companies.Command.print_count')
     def test_handler_with_options(self, print_count, save_suppliers, drop_all, print_):
         print_count.return_value = 0
         self.command.handle(dataset='companies.xz', drop=True)
